Remove description of CategoryData
When first installing I created a CategoryData type as the readme said "CategoryData looks like this"  which made me think I needed to set it up as part of installation as it immediately followed a description of code that needed to be created. I thought I was the only one who made this mistake but seems I am not the only one: [https://world.episerver.com/forum/developer-forum/-Episerver-75-CMS/Thread-Container/2020/3/geta-epicategories---cant-add-categories/]. So suggesting to remove the description from readme and update the comments accordingly.